Mais debug

+++ parte2/instr.h
@@ -62,8 +62,8 @@ typedef enum {//this new
 /* Célula */
 typedef struct {
   Terreno terreno;
-  short int cristais;
-  short int ocupado;
+  int cristais;
+  int ocupado;
   int pos[2];
   int base;//numero vai ser o time; 0 = nao tem time
 } Celula;
